% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_agg.R
\name{agg_by_date}
\alias{agg_by_date}
\title{Specify a date binning for aggregation}
\usage{
agg_by_date(query, field, interval = calendar_interval("1d"))
}
\arguments{
\item{query}{a \code{\link[=query_agg]{query_agg()}} object}

\item{field}{field name (see \code{\link[=queryable_fields]{queryable_fields()}} for all possibilities)}

\item{interval}{binning interval (see \code{\link[=calendar_interval]{calendar_interval()}} or \code{\link[=fixed_interval]{fixed_interval()}}) - default is 1-day calendar interval}
}
\description{
Specify a date binning for aggregation
}
