% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_agg.R
\name{agg_by_field}
\alias{agg_by_field}
\title{Specify a field to aggregate by}
\usage{
agg_by_field(query, field)
}
\arguments{
\item{query}{a \code{\link[=query_agg]{query_agg()}} object}

\item{field}{field name (see \code{\link[=queryable_fields]{queryable_fields()}} for all possibilities)}
}
\description{
Specify a field to aggregate by
}
