% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{calendar_interval}
\alias{calendar_interval}
\title{Specify a calendar interval for date binning aggregation}
\usage{
calendar_interval(unit = "1d")
}
\arguments{
\item{unit}{one of "1m" (minute), "1h" (hour), "1d" (day), "1w" (week), "1M" (month), "1q" (quarter), "1y" (year)}
}
\description{
Specify a calendar interval for date binning aggregation
}
