% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{es_connect}
\alias{es_connect}
\title{Set connection details to Elasticsearch engine}
\usage{
es_connect(
  host = "127.0.0.1",
  port = 9200,
  primary_index = NULL,
  path = NULL,
  transport_schema = "http",
  user = NULL,
  pwd = NULL,
  headers = NULL,
  cainfo = NULL,
  force = FALSE,
  errors = "simple",
  warn = TRUE,
  ...
)
}
\arguments{
\item{host}{(character) The base host, defaults to \verb{127.0.0.1}}

\item{port}{(character) port to connect to, defaults to \code{9200}
(optional)}

\item{primary_index}{the name of the primary index to use for querying (will be used as a fallback if index is not explicitly specified in the query)}

\item{path}{(character) context path that is appended to the end of the
url. Default: \code{NULL}, ignored}

\item{transport_schema}{(character) http or https. Default: \code{http}}

\item{user}{(character) User name, if required for the connection. You
can specify,  but ignored for now.}

\item{pwd}{(character) Password, if required for the connection. You
can specify, but ignored for now.}

\item{headers}{named list of headers. These headers are used in all requests}

\item{cainfo}{(character) path to a crt bundle, passed to curl option
\code{cainfo}}

\item{force}{(logical) Force re-load of connection details.
Default: \code{FALSE}}

\item{errors}{(character) One of simple (Default) or complete. Simple gives
http code and  error message on an error, while complete gives both http
code and error message,  and stack trace, if available.}

\item{warn}{(logical) whether to throw warnings from the Elasticsearch
server when provided. Pulls warnings from response headers when given.
default: \code{TRUE}. To turn these off, you can set \code{warn=FALSE} or
wrap function calls in \code{\link[=suppressWarnings]{suppressWarnings()}}. You can also see warnings in
headers by using curl verbose.}

\item{...}{additional curl options to be passed in ALL http requests}
}
\description{
Set connection details to Elasticsearch engine
}
\details{
See \code{\link[elastic:connect]{elastic::connect()}} for more details.
}
\examples{
\dontrun{
con <- es_connect(host = "localhost", port = 9200)
}
}
