% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_match}
\alias{filter_match}
\title{Specify a match filter}
\usage{
filter_match(query, field, match)
}
\arguments{
\item{query}{a \code{\link[=query_agg]{query_agg()}} or \code{\link[=query_fetch]{query_fetch()}} object}

\item{field}{field name (see \code{\link[=queryable_fields]{queryable_fields()}} for all possibilities)}

\item{match}{a string to match}
}
\description{
Specify a match filter
}
