% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_range}
\alias{filter_range}
\title{Specify a range filter}
\usage{
filter_range(query, field, from = NULL, to = NULL)
}
\arguments{
\item{query}{a \code{\link[=query_agg]{query_agg()}} or \code{\link[=query_fetch]{query_fetch()}} object}

\item{field}{field name (see \code{\link[=queryable_fields]{queryable_fields()}} for all possibilities)}

\item{from}{the lower value of the range}

\item{to}{the upper value of the range}
}
\description{
Specify a range filter
}
