% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_regexp}
\alias{filter_regexp}
\title{Specify a regexp filter}
\usage{
filter_regexp(
  query,
  field,
  regexp,
  bool = c("must", "filter", "must_not", "should")
)
}
\arguments{
\item{query}{a \code{\link[=query_agg]{query_agg()}} or \code{\link[=query_fetch]{query_fetch()}} object}

\item{field}{field name (see \code{\link[=queryable_fields]{queryable_fields()}} for all possibilities)}

\item{regexp}{a regular expression string to match}

\item{bool}{one of "must", "filter", "must_not", "should" (see \href{https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-bool-query.html}{here})}
}
\description{
Specify a regexp filter
}
