% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{fixed_interval}
\alias{fixed_interval}
\title{Specify a fixed time interval for date binning aggregation}
\usage{
fixed_interval(unit = "1d")
}
\arguments{
\item{unit}{a positive integer followed by one of "ms" (millisecond), "s" (second), "m" (minute), "h" (hour), "d" (day)}
}
\description{
Specify a fixed time interval for date binning aggregation
}
