% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_fetch.R
\name{query_fetch}
\alias{query_fetch}
\title{Initialize a document fetch query}
\usage{
query_fetch(
  con,
  index = NULL,
  path = NULL,
  size = 10000,
  time_scroll = "5m",
  max = 0,
  format = c("list", "json"),
  check_fields = TRUE
)
}
\arguments{
\item{con}{an \code{\link[=es_connect]{es_connect()}} object}

\item{index}{the name of the index to query (if not specified, will fall back on 'primary_index' provided to \code{\link[=es_connect]{es_connect()}})}

\item{path}{an optional directory to write documents to as they are fetched}

\item{size}{the number of documents to fetch in each batch (max is 10000)}

\item{time_scroll}{specify how long a consistent view of the index should be maintained for scrolled search, e.g. "1m"}

\item{max}{maximum number of documents to fetch}

\item{format}{format of returned output - either "list" or "json"}

\item{check_fields}{should field names be checked when the query is
augmented?}
}
\description{
Initialize a document fetch query
}
