% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_str.R
\name{query_str}
\alias{query_str}
\title{Initialize a query with a specified query string}
\usage{
query_str(
  con,
  index = NULL,
  str,
  path = NULL,
  time_scroll = "5m",
  max = 0,
  type = c("unknown", "fetch", "agg"),
  format = c("list", "json"),
  check_fields = TRUE
)
}
\arguments{
\item{con}{an \code{\link[=es_connect]{es_connect()}} object}

\item{index}{the name of the index to query (if not specified, will
fall back on 'primary_index' provided to \code{\link[=es_connect]{es_connect()}})}

\item{str}{query string}

\item{path}{an optional directory to write documents to as they are fetched}

\item{time_scroll}{specify how long a consistent view of the index should
be maintained for scrolled search, e.g. "1m" (used when type=="fetch"}

\item{max}{maximum number of documents to fetch (used when type=="fetch")}

\item{type}{the type of query defined in the string - one of "unknown",
"fetch", or "agg"}

\item{format}{format of returned output - either "list" or "json"}

\item{check_fields}{should field names be checked when the query is
augmented?}
}
\description{
Initialize a query with a specified query string
}
