% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{queryable_fields}
\alias{queryable_fields}
\title{Get list of all queryable fields}
\usage{
queryable_fields(con)
}
\arguments{
\item{con}{an \code{\link[=es_connect]{es_connect()}} object}
}
\description{
Get list of all queryable fields
}
