% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{run}
\alias{run}
\title{Execute a query}
\usage{
run(query)
}
\arguments{
\item{query}{a \code{\link[=query_agg]{query_agg()}}, \code{\link[=query_fetch]{query_fetch()}}, or \code{\link[=query_str]{query_str()}} object}
}
\description{
Execute a query
}
