% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_ops.R
\name{select_fields}
\alias{select_fields}
\title{Specify fields to return}
\usage{
select_fields(query, fields)
}
\arguments{
\item{query}{a \code{\link[=query_fetch]{query_fetch()}} object}

\item{fields}{field name or vector of field names (see \code{\link[=selectable_fields]{selectable_fields()}} for all possibilities)}
}
\description{
Specify fields to return
}
