% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectable_fields}
\alias{selectable_fields}
\title{Get list of all selectable fields}
\usage{
selectable_fields(con)
}
\arguments{
\item{con}{an \code{\link[=es_connect]{es_connect()}} object}
}
\description{
Get list of all selectable fields
}
