% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_ops.R
\name{sort_docs}
\alias{sort_docs}
\title{Specify sort order of returned documents}
\usage{
sort_docs(query, fields)
}
\arguments{
\item{query}{a \code{\link[=query_fetch]{query_fetch()}} object}

\item{fields}{field name or vector of field names (see \code{\link[=queryable_fields]{queryable_fields()}} for all possibilities)}
}
\description{
Specify sort order of returned documents
}
