% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_terms}
\alias{filter_terms}
\title{Specify a terms filter}
\usage{
filter_terms(
  query,
  field,
  terms,
  bool = c("must", "filter", "must_not", "should")
)
}
\arguments{
\item{query}{a \code{\link[=query_agg]{query_agg()}} or \code{\link[=query_fetch]{query_fetch()}} object}

\item{field}{field name (see \code{\link[=queryable_fields]{queryable_fields()}} for all possibilities)}

\item{terms}{a string or vector of strings to exact match}

\item{bool}{one of "must", "filter", "must_not", "should" (see \href{https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-bool-query.html}{here})}
}
\description{
Specify a terms filter
}
