% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{query_fetch}
\alias{query_fetch}
\title{Initialize an document fetch query}
\usage{
query_fetch(
  con,
  index = NULL,
  path = NULL,
  size = 10000,
  time_scroll = "5m",
  max = 0
)
}
\arguments{
\item{con}{an \code{\link[=es_connect]{es_connect()}} object}

\item{index}{the name of the index to query (if not specified, will fall back on 'primary_index' provided to \code{\link[=es_connect]{es_connect()}})}

\item{path}{an optional directory to write docuents to as they are fetched}

\item{size}{the number of documents to fetch in each batch (max is 10000)}

\item{time_scroll}{specify how long a consistent view of the index should be maintained for scrolled search, e.g. "1m"}

\item{max}{maximum number of documents to fetch}
}
\description{
Initialize an document fetch query
}
